requirements.txt
Django==5.0.6
mysqlclient==2.2.4
requests==2.31.0
djangorestframework==3.15.1